# api/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  description: |
    RESTful API for multi-tenant blog platform.

    ## Authentication

    All endpoints (except `/api/health`) require API key authentication.

    Include your API key in one of these headers:
    - `Authorization: Bearer <your-api-key>`
    - `X-API-Key: <your-api-key>`

    ## API Key Types

    - **Admin keys** (`sa_`): Full access to all resources
    - **User keys** (`sk_`): Access to owned/member sites
    - **Site keys** (`ss_`): Access limited to specific site

    ## Scopes

    - `read`: Read-only access
    - `write`: Create and update resources
    - `delete`: Delete resources
    - `admin`: Full administrative access
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in X-API-Key header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error details
      required:
        - error
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        key_prefix:
          type: string
          description: First characters of key for identification
        key_type:
          type: string
          enum: [user, site, admin]
        user_id:
          type: string
          format: uuid
          nullable: true
        site_id:
          type: string
          format: uuid
          nullable: true
        scopes:
          type: array
          items:
            type: string
            enum: [read, write, delete, admin]
        rate_limit_per_minute:
          type: integer
        rate_limit_per_day:
          type: integer
        is_active:
          type: boolean
        last_used_at:
          type: string
          format: date-time
          nullable: true
        usage_count:
          type: integer
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ApiKeyCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        key_type:
          type: string
          enum: [user, site, admin]
          default: user
        user_id:
          type: string
          format: uuid
          description: Required for admin keys creating user/site keys
        site_id:
          type: string
          format: uuid
          description: Required for site keys
        scopes:
          type: array
          items:
            type: string
            enum: [read, write, delete, admin]
          default: [read]
        rate_limit_per_minute:
          type: integer
          default: 60
        rate_limit_per_day:
          type: integer
          default: 10000
        expires_at:
          type: string
          format: date-time
        allowed_ips:
          type: array
          items:
            type: string
        allowed_origins:
          type: array
          items:
            type: string
        metadata:
          type: object
    ApiKeyCreated:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          description: Full API key - only shown once at creation
        key_prefix:
          type: string
        name:
          type: string
        key_type:
          type: string
          enum: [user, site, admin]
    ApiKeyUsageStats:
      type: object
      properties:
        totalRequests:
          type: integer
        successfulRequests:
          type: integer
        failedRequests:
          type: integer
        avgResponseTime:
          type: number
        topEndpoints:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
              count:
                type: integer
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/api-keys:
    get:
      tags:
        - API Keys
      summary: List API keys
      description: List API keys (users see their own, admins can see all)
      parameters:
        - name: user_id
          in: query
          description: Filter by user (admin only)
          schema:
            type: string
            format: uuid
        - name: key_type
          in: query
          schema:
            type: string
            enum: [user, site, admin]
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  count:
                    type: integer

    post:
      tags:
        - API Keys
      summary: Create API key
      description: |
        Create a new API key. The full key is only returned once at creation.

        - Users can create user keys for themselves
        - Users can create site keys for sites they own
        - Only admin keys can create admin keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ApiKeyCreated'
                  message:
                    type: string
                    example: Save this API key securely - it will not be shown again
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/api-keys/{keyId}:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - API Keys
      summary: Get API key
      description: Get details of a specific API key
      responses:
        '200':
          description: API key details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ApiKey'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - API Keys
      summary: Update API key
      description: Update API key metadata (cannot change type or scopes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                rate_limit_per_minute:
                  type: integer
                rate_limit_per_day:
                  type: integer
                allowed_ips:
                  type: array
                  items:
                    type: string
                allowed_origins:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
      responses:
        '200':
          description: API key updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ApiKey'
  /api/api-keys/{keyId}/revoke:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - API Keys
      summary: Revoke API key
      description: Revoke an API key (cannot be undone)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for revocation
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/api-keys/{keyId}/usage:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - API Keys
      summary: Get API key usage
      description: Get usage statistics for an API key
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ApiKeyUsageStats'
  /api/api-keys/{keyId}/site-access:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - API Keys
      summary: Grant site access
      description: Grant an API key access to a specific site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - site_id
                - scopes
              properties:
                site_id:
                  type: string
                  format: uuid
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [read, write, delete]
      responses:
        '200':
          description: Site access granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/api-keys/{keyId}/site-access/{siteId}:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: siteId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      tags:
        - API Keys
      summary: Revoke site access
      description: Revoke an API key's access to a specific site
      responses:
        '200':
          description: Site access revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string